Fix discovery logic for Haiku Lights
Fix error discovery logic that treated a rediscovered Haiku Light as a new device. Caused known device list to grow and a new connection to the fan each time it was discovered.
Removed some unhelpful discovery debug messages.
Cleanup the usage of the word fan when device should be used.